Updating scripts to accept files and folders

+++ scripts/crypto.js
@@ -1,26 +1,44 @@
 import AdmZip from 'adm-zip'
 import fs from 'fs'
 import { encryptZipFile, decryptZipFile } from './zip'
+import path from 'path'
 
-export async function encryptFile(folderPath, password) {
-  const zipFilePath = `${folderPath}.zip`
+export async function encryptFile(inputPath, password) {
+  const stats = fs.statSync(inputPath)
+  const isDirectory = stats.isDirectory()
 
   const zip = new AdmZip()
-  zip.addLocalFolder(folderPath)
+  const zipFilePath = `${inputPath}.zip`
+
+  if (isDirectory) {
+    zip.addLocalFolder(inputPath)
+  } else {
+    const fileData = fs.readFileSync(inputPath)
+    const relativeFilePath = inputPath.split('/').pop()
+    zip.addFile(relativeFilePath, fileData)
+  }
+
   zip.writeZip(zipFilePath)
 
   encryptZipFile(zipFilePath, password)
 }
 
-export async function decryptFile(encryptedFolderPath, password) {
-  const encryptedFilePath = `${encryptedFolderPath}.chameleon`
+export async function decryptFile(inputPath, password) {
+  const stats = fs.statSync(inputPath)
+  const isDirectory = stats.isDirectory()
 
-  const decryptedFilePath = decryptZipFile(encryptedFilePath, password)
+  let decryptedFolder
+  if (isDirectory) {
+    decryptedFolder = inputPath.replace('.chameleon', '')
+    const zip = new AdmZip(inputPath)
+    zip.extractAllTo(decryptedFolder, true)
+  } else {
+    const decryptedFilePath = decryptZipFile(inputPath, password)
+    decryptedFolder = path.dirname(inputPath)
+    const zip = new AdmZip(decryptedFilePath)
+    zip.extractAllTo(decryptedFolder, true)
+    fs.unlinkSync(decryptedFilePath)
+  }
 
-  const decryptedFolder = decryptedFilePath.replace('.zip', '')
-
-  const zip = new AdmZip(decryptedFilePath)
-  zip.extractAllTo(decryptedFolder, true)
-
-  fs.unlinkSync(decryptedFilePath)
+  return decryptedFolder
 }
